+++ src/commands/mod.rs
@@ -2,6 +2,7 @@ pub mod cut;
 pub mod delete;
 pub mod fade;
 pub mod gain;
+pub mod marker;
 pub mod normalize;
 pub mod paste;
 pub mod resample;